+++ frontend/src/__tests__/emotion-education-page.test.tsx
@@ -1,5 +1,5 @@
 /// <reference types="vitest/globals" />
-import NonCognitivePage from '@/app/(authed)/app/non-cognitive/page';
+import EmotionEducationPage from '../app/(authed)/app/emotion-education/page';
 import { useAuth } from '@/contexts/AuthContext';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { useRouter } from 'next/navigation';
@@ -15,7 +15,7 @@ vi.mock('@/components/ui', () => ({
   HamburgerMenu: () => <div>Menu</div>,
 }));
 
-describe('NonCognitivePage', () => {
+describe('EmotionEducationPage', () => {
   const mockPush = vi.fn();
   const mockRouter = { push: mockPush };
 
@@ -29,22 +29,22 @@ describe('NonCognitivePage', () => {
     isLoading = false,
   }: { user?: { id: string } | null; isLoading?: boolean } = {}) => {
     (useAuth as any).mockReturnValue({ user, isLoading });
-    render(<NonCognitivePage />);
+    render(<EmotionEducationPage />);
   };
 
   it('正常にレンダリングされる', () => {
     setup();
-    expect(screen.getByText('非認知能力について')).toBeInTheDocument();
+    expect(screen.getByText('感情教育について')).toBeInTheDocument();
     expect(
-      screen.getByRole('heading', { name: '非認知能力とは？' }),
+      screen.getByRole('heading', { name: '感情教育とは？' }),
     ).toBeInTheDocument();
   });
 
   it('タブ切り替えが動作する', () => {
     setup();
-    fireEvent.click(screen.getByText('非認知能力の重要性'));
+    fireEvent.click(screen.getByText('なぜ大切なのか？'));
     expect(
-      screen.getByRole('heading', { name: '非認知能力の重要性' }),
+      screen.getByRole('heading', { name: 'なぜ大切なのか？' }),
     ).toBeInTheDocument();
   });
