+++ frontend/src/__tests__/tips-page.test.tsx
@@ -1,5 +1,5 @@
 /// <reference types="vitest/globals" />
-import UsagePage from '@/app/(authed)/app/usage/page';
+import TipsPage from '../app/(authed)/app/tips/page';
 import { useAuth } from '@/contexts/AuthContext';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { useRouter } from 'next/navigation';
@@ -15,7 +15,7 @@ vi.mock('@/components/ui', () => ({
   HamburgerMenu: () => <div>Menu</div>,
 }));
 
-describe('UsagePage', () => {
+describe('TipsPage', () => {
   const mockPush = vi.fn();
   const mockRouter = { push: mockPush };
 
@@ -24,36 +24,30 @@ describe('UsagePage', () => {
     (useRouter as any).mockReturnValue(mockRouter);
   });
 
-  // 共通 setup
   const setup = ({
     user = { id: '1' },
     isLoading = false,
   }: { user?: { id: string } | null; isLoading?: boolean } = {}) => {
     (useAuth as any).mockReturnValue({ user, isLoading });
-    render(<UsagePage />);
+    render(<TipsPage />);
   };
 
   it('正常にレンダリングされる', () => {
     setup();
-    expect(screen.getByText('使い方')).toBeInTheDocument();
-    expect(screen.getByText('感情記録を始める')).toBeInTheDocument();
-  });
-
-  it('タブ切り替えが動作する', () => {
-    setup();
-    // ボタンの「毎日の流れ」をクリック
-    const tabButton = screen.getByRole('button', { name: '毎日の流れ' });
-    fireEvent.click(tabButton);
-    // h2の「毎日の流れ」が表示されることを確認
+    expect(screen.getByText('保護者向けTips')).toBeInTheDocument();
     expect(
-      screen.getByRole('heading', { name: '毎日の流れ' }),
+      screen.getByText(/感情に名前をつけてみましょう/),
     ).toBeInTheDocument();
   });
 
-  it('ボタンクリックが動作する', () => {
+  it('戻るボタンが動作する', () => {
     setup();
     fireEvent.click(screen.getByText('← もどる'));
     expect(mockPush).toHaveBeenCalledWith('/app');
+  });
+
+  it('感情記録ボタンが動作する', () => {
+    setup();
     fireEvent.click(screen.getByText('感情記録を始める'));
     expect(mockPush).toHaveBeenCalledWith('/app/emotion-selection');
   });
