+++ frontend/src/__tests__/usage-page.test.tsx
@@ -1,5 +1,5 @@
 /// <reference types="vitest/globals" />
-import FAQPage from '@/app/(authed)/app/faq/page';
+import UsagePage from '../app/(authed)/app/usage/page';
 import { useAuth } from '@/contexts/AuthContext';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { useRouter } from 'next/navigation';
@@ -15,7 +15,7 @@ vi.mock('@/components/ui', () => ({
   HamburgerMenu: () => <div>Menu</div>,
 }));
 
-describe('FAQPage', () => {
+describe('UsagePage', () => {
   const mockPush = vi.fn();
   const mockRouter = { push: mockPush };
 
@@ -24,30 +24,36 @@ describe('FAQPage', () => {
     (useRouter as any).mockReturnValue(mockRouter);
   });
 
+  // 共通 setup
   const setup = ({
     user = { id: '1' },
     isLoading = false,
   }: { user?: { id: string } | null; isLoading?: boolean } = {}) => {
     (useAuth as any).mockReturnValue({ user, isLoading });
-    render(<FAQPage />);
+    render(<UsagePage />);
   };
 
   it('正常にレンダリングされる', () => {
     setup();
-    expect(screen.getByText('よくある質問（FAQ）')).toBeInTheDocument();
+    expect(screen.getByText('使い方')).toBeInTheDocument();
+    expect(screen.getByText('感情記録を始める')).toBeInTheDocument();
+  });
+
+  it('タブ切り替えが動作する', () => {
+    setup();
+    // ボタンの「毎日の流れ」をクリック
+    const tabButton = screen.getByRole('button', { name: '毎日の流れ' });
+    fireEvent.click(tabButton);
+    // h2の「毎日の流れ」が表示されることを確認
     expect(
-      screen.getByText(/このアプリは何歳から使えますか？/),
+      screen.getByRole('heading', { name: '毎日の流れ' }),
     ).toBeInTheDocument();
   });
 
-  it('戻るボタンが動作する', () => {
+  it('ボタンクリックが動作する', () => {
     setup();
     fireEvent.click(screen.getByText('← もどる'));
     expect(mockPush).toHaveBeenCalledWith('/app');
-  });
-
-  it('感情記録ボタンが動作する', () => {
-    setup();
     fireEvent.click(screen.getByText('感情記録を始める'));
     expect(mockPush).toHaveBeenCalledWith('/app/emotion-selection');
   });
